refactor(tests): extract engine_indexing.rs (7 tests)
Move batch_indexed_* + engine_* + lto_eliminates_fault_injection from
mod.rs into engine/tests/engine_indexing.rs.

Open-at-execution decision: lto_eliminates_fault_injection HOME =
engine_indexing.rs (not standalone lto.rs). Reason: the assertion is
about engine-indexing behaviour under LTO, and its natural neighbour is
engine_fallback_stale_on_update_failure (the non-LTO fault-injection
companion). Single-test lto.rs would fragment the pair.

Tests moved (now at engine::tests::engine_indexing::*):
  batch_indexed_docs_have_code_spans
  batch_indexed_docs_preserve_frontmatter
  engine_index_creates_persistent_engines
  engine_fallback_stale_on_update_failure
  lto_eliminates_fault_injection
  engine_cleanup_on_root_removal
  engine_frontmatter_preserved

Tests: 135 passed / 0 failed / 1 ignored (baseline unchanged).

marky-n1h step 7.
